+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * #4044: Ensure agent uses bi-directional AMQP idle-time-out for all connections
 * #4076: Avoid memory leak if connected router is disconnected
+* #4117: Add support for configuring console to use impersonation
 
 
 ## 0.31.0